requirements.txt
# AstroLife Explorer Requirements
transformers==4.30.2
torch==2.0.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
tqdm==4.65.0
flask==2.3.2
python-dotenv==1.0.0
